CR fixes 2
