+++ runtest-moduleapi
@@ -41,5 +41,5 @@ $TCLSH tests/test_helper.tcl \
 --single unit/moduleapi/stream \
 --single unit/moduleapi/datatype2 \
 --single unit/moduleapi/aclcheck \
---single unit/moduleapi/container \
+--single unit/moduleapi/subcommands \
 "${@}"
